refactor: Student Profile 분리
