+++ src/main/java/com/chukchuk/haksa/domain/academic/record/controller/AcademicRecordController.java
@@ -50,7 +50,6 @@ public class AcademicRecordController {
         UUID userId = UUID.fromString(userDetails.getUsername());
         StudentAcademicRecordDto.AcademicSummaryDto academicSummary = academicRecordService.getAcademicSummary(userId);
 
-
         return ResponseEntity.ok(academicSummary);
     }
