+++ src/main/java/com/chukchuk/haksa/domain/academic/record/dto/StudentAcademicRecordDto.java
@@ -12,6 +12,8 @@ public class StudentAcademicRecordDto {
             @Schema(description = "총 취득 학점", example = "120") Integer totalEarnedCredits,
             @Schema(description = "누적 GPA", example = "3.76") BigDecimal cumulativeGpa,
             @Schema(description = "전체 백분위", example = "87.5") BigDecimal percentile
+//            @Schema(description = "필요 졸업 학점", example = "130") Integer
+//             requiredCredits
     ) {
         public static AcademicSummaryDto from(StudentAcademicRecord studentAcademicRecord) {
             return new AcademicSummaryDto(
