+++ src/main/java/com/chukchuk/haksa/domain/student/dto/StudentDto.java
@@ -2,8 +2,10 @@ package com.chukchuk.haksa.domain.student.dto;
 
 import com.chukchuk.haksa.domain.student.model.Student;
 import com.chukchuk.haksa.domain.student.model.StudentStatus;
+import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.Instant;
+import java.util.Objects;
 
 public class StudentDto {
 
@@ -27,6 +29,33 @@ public class StudentDto {
                     student.getAcademicInfo().getStatus(),
                     student.getAcademicInfo().getCompletedSemesters(),
                     student.getUpdatedAt()
+            );
+        }
+    }
+
+    @Schema(description = "학생 프로필 정보")
+    public record Profile(
+            @Schema(description = "이름") String name,
+            @Schema(description = "학번") String studentCode,
+            @Schema(description = "학과 이름") String departmentName,
+            @Schema(description = "전공 이름") String majorName,
+            @Schema(description = "학년") int gradeLevel,
+            @Schema(description = "현재 학기") int currentSemester,
+            @Schema(description = "재학 상태") String status,
+            @Schema(description = "마지막 업데이트 일시") String lastUpdatedAt,
+            @Schema(description = "학사 정보 마지막 연동 일시") String lastSyncedAt
+    ) {
+        public static Profile from(StudentDto.StudentInfoDto studentInfoDto, int currentSemester, String lastSyncedAt) {
+            return new Profile(
+                    Objects.requireNonNullElse(studentInfoDto.name(), ""),
+                    Objects.requireNonNullElse(studentInfoDto.studentCode(), ""),
+                    Objects.requireNonNullElse(studentInfoDto.departmentName(), ""),
+                    Objects.requireNonNullElse(studentInfoDto.majorName(), ""),
+                    studentInfoDto.gradeLevel() != null ? studentInfoDto.gradeLevel() : 0,
+                    currentSemester,
+                    Objects.requireNonNullElse(studentInfoDto.status().toString(), ""),
+                    studentInfoDto.updatedAt().toString(),
+                    lastSyncedAt
             );
         }
     }
