+++ src/main/java/com/chukchuk/haksa/domain/student/service/StudentService.java
@@ -3,6 +3,7 @@ package com.chukchuk.haksa.domain.student.service;
 import com.chukchuk.haksa.domain.student.dto.StudentDto;
 import com.chukchuk.haksa.domain.student.model.Student;
 import com.chukchuk.haksa.domain.student.repository.StudentRepository;
+import com.chukchuk.haksa.domain.user.model.User;
 import com.chukchuk.haksa.domain.user.service.UserService;
 import com.chukchuk.haksa.global.exception.DataNotFoundException;
 import lombok.RequiredArgsConstructor;
@@ -19,17 +20,19 @@ public class StudentService {
     private final StudentRepository studentRepository;
     private final UserService userService;
 
-    public StudentDto.StudentInfoDto getStudentInfo(UUID userId) {
-
-        // 학생 조회
-        Student student = getStudentById(userId);
-
-        return StudentDto.StudentInfoDto.from(student);
-    }
-
     public Student getStudentById(UUID studentId) {
         return studentRepository.findById(studentId)
                 .orElseThrow(() -> new DataNotFoundException("해당 학생이 존재하지 않습니다."));
+    }
+
+    public StudentDto.Profile getStudentProfile(UUID userId) {
+        StudentDto.StudentInfoDto studentInfo = getStudentInfo(userId);
+        int currentSemester = getCurrentSemester(studentInfo.gradeLevel(), studentInfo.completedSemesters());
+
+        User user = userService.getUserById(userId);
+        String lastSyncedAt = user.getLastSyncedAt() != null ? user.getLastSyncedAt().toString() : "";
+
+        return StudentDto.Profile.from(studentInfo, currentSemester, lastSyncedAt);
     }
 
     @Transactional
@@ -44,5 +47,35 @@ public class StudentService {
 
         student.setTargetGpa(targetGpa);
         studentRepository.save(student);
+    }
+
+    private StudentDto.StudentInfoDto getStudentInfo(UUID userId) {
+        // 학생 조회
+        Student student = getStudentById(userId);
+
+        return StudentDto.StudentInfoDto.from(student);
+    }
+
+    private static int getCurrentSemester(Integer gradeLevel, Integer completedSemesters) {
+
+        int safeGradeLevel = (gradeLevel != null) ? gradeLevel : 0;
+        int safeCompletedSemesters = (completedSemesters != null) ? completedSemesters : 0;
+
+        // 1️⃣ 해당 학년 이전에 완료해야 하는 최소 학기 수
+        int expectedCompleted = (safeGradeLevel - 1) * 2;
+
+        // 2️⃣ 만약 completedSemesters가 예상보다 낮으면 최소값으로 간주 (방학 중 업데이트가 안된 경우 대비)
+        int effectiveCompleted = Math.max(safeCompletedSemesters, expectedCompleted);
+
+        // 3️⃣ 현재 학기 계산
+        int currentSemester = effectiveCompleted - expectedCompleted + 1;
+
+        // 4️⃣ currentSemester는 1 또는 2만 가능하도록 보정
+        if (currentSemester < 1) {
+            return 1;
+        } else if (currentSemester > 2) {
+            return 2;
+        }
+        return currentSemester;
     }
 }
